Move BlockSlots class to DPOS domain
