+++ framework/src/modules/chain/dpos/index.js
@@ -1,5 +1,23 @@
+/*
+ * Copyright © 2019 Lisk Foundation
+ *
+ * See the LICENSE file at the top-level directory of this distribution
+ * for licensing information.
+ *
+ * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
+ * no part of this software, including this file, may be copied, modified,
+ * propagated, or distributed except according to the terms contained in the
+ * LICENSE file.
+ *
+ * Removal or modification of this copyright notice is prohibited.
+ */
+
+'use strict';
+
+const { BlockSlots } = require('./block_slots');
 const Dpos = require('./dpos');
 
 module.exports = {
+	BlockSlots,
 	Dpos,
 };
