+++ framework/test/jest/unit/specs/modules/chain/blocks/fork_choice_rule.spec.js
@@ -6,9 +6,7 @@ const {
 	isTieBreak,
 	isValidBlock,
 } = require('../../../../../../../src/modules/chain/blocks/fork_choice_rule');
-const {
-	BlockSlots,
-} = require('../../../../../../../src/modules/chain/blocks/block_slots');
+const { BlockSlots } = require('../../../../../../../src/modules/chain/dpos');
 
 const EPOCH_TIME = new Date(Date.UTC(2016, 4, 24, 17, 0, 0, 0)).toISOString();
 const BLOCK_TIME = 10;
