+++ framework/test/mocha/integration/blocks/process/on_receive_block.js
@@ -23,7 +23,7 @@ const {
 	getPrivateAndPublicKeyBytesFromPassphrase,
 } = require('@liskhq/lisk-cryptography');
 const accountFixtures = require('../../../fixtures/accounts');
-const { BlockSlots } = require('../../../../../src/modules/chain/blocks');
+const { BlockSlots } = require('../../../../../src/modules/chain/dpos');
 const blocksUtils = require('../../../../../src/modules/chain/blocks/block');
 const blockVersion = require('../../../../../src/modules/chain/blocks/block_version');
 const genesisDelegates = require('../../../data/genesis_delegates.json')
