+++ framework/test/mocha/unit/modules/chain/dpos/block_slots.js
@@ -14,9 +14,7 @@
 
 'use strict';
 
-const {
-	BlockSlots,
-} = require('../../../../../../src/modules/chain/blocks/block_slots');
+const { BlockSlots } = require('../../../../../../src/modules/chain/dpos');
 
 describe('BlockSlots', () => {
 	let slots;
